+++ android/src/car/superfun/game/AndroidLauncher.java
@@ -20,7 +20,7 @@ import com.instacart.library.truetime.TrueTime;
 import java.io.IOException;
 import java.util.ArrayList;
 
-import car.superfun.game.car.OpponentCarController;
+import car.superfun.game.cars.OpponentCarController;
 import car.superfun.game.googleGamePlayServices.Communicator;
 import car.superfun.game.googleGamePlayServices.GoogleGameServices;
 import car.superfun.game.googleGamePlayServices.SetUpGame;
