+++ android/src/car/superfun/game/gameModes/gladiatorMode/LocalGladiatorCar.java
@@ -5,8 +5,8 @@ import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.physics.box2d.World;
 
 import car.superfun.game.GlobalVariables;
-import car.superfun.game.car.LocalCarController;
-import car.superfun.game.car.LocalCar;
+import car.superfun.game.cars.LocalCarController;
+import car.superfun.game.cars.LocalCar;
 
 import static java.lang.Math.abs;
